+++ designer/server/src/routes/forms/editor-v2/condition.js
@@ -1,12 +1,7 @@
 import { randomUUID } from 'crypto'
 import Stream from 'node:stream'
 
-import {
-  FormDefinitionError,
-  conditionDataSchemaV2,
-  conditionWrapperSchemaV2,
-  slugSchema
-} from '@defra/forms-model'
+import { FormDefinitionError, slugSchema } from '@defra/forms-model'
 import Boom from '@hapi/boom'
 import { StatusCodes } from 'http-status-codes'
 import Joi from 'joi'
@@ -26,6 +21,7 @@ import {
   getConditionSessionState
 } from '~/src/lib/session-helper.js'
 import { CHANGES_SAVED_SUCCESSFULLY } from '~/src/models/forms/editor-v2/common.js'
+import { conditionWrapperSchema } from '~/src/models/forms/editor-v2/condition-helper.js'
 import * as viewModel from '~/src/models/forms/editor-v2/condition.js'
 import { editorFormPath, editorv2Path } from '~/src/models/links.js'
 import {
@@ -42,62 +38,6 @@ const notificationKey = sessionNames.successNotification
 
 const idSchema = Joi.string().uuid().allow('new').required()
 const stateIdSchema = Joi.string().optional()
-const componentIdSchema = conditionDataSchemaV2.extract('componentId')
-const operatorSchema = conditionDataSchemaV2.extract('operator')
-const valueSchema = conditionDataSchemaV2.extract('value')
-const typeSchema = conditionDataSchemaV2.extract('type')
-
-/**
- * @type {Joi.ObjectSchema<ConditionWrapperPayload>}
- * Custom condition wrapper payload schema that
- * only allows conditions, not condition references.
- *
- * There is a dependency chain in the validation of each condition item:
- * - Condition operator is only required once a componentId has been selected
- * - Condition value is only required once a valid operator has been selected
- * Given this, we don't want to surface errors to the user for operator and
- * value before their dependent fields are valid hence the use of `joi.when` below
- */
-const conditionWrapperSchema = conditionWrapperSchemaV2.keys({
-  coordinator: conditionWrapperSchemaV2.extract('coordinator').messages({
-    'any.required': 'Choose how you want to combine conditions'
-  }),
-  items: Joi.array().items(
-    conditionDataSchemaV2.keys({
-      componentId: componentIdSchema.messages({
-        '*': 'Select a question'
-      }),
-      operator: operatorSchema
-        .when('componentId', {
-          not: componentIdSchema,
-          then: Joi.optional() // Only validate the operator if the componentId is valid
-        })
-        .messages({
-          '*': 'Select a condition type'
-        }),
-      type: typeSchema
-        .when('operator', {
-          not: operatorSchema,
-          then: Joi.optional() // Only validate the value if the operator is valid
-        })
-        .messages({
-          '*': 'Enter a condition value type'
-        }),
-      value: valueSchema
-        .when('operator', {
-          not: operatorSchema,
-          then: Joi.optional() // Only validate the value if the operator is valid
-        })
-        .messages({
-          '*': 'Enter a condition value',
-          'date.format': 'Enter a condition value in the correct format'
-        })
-    })
-  ),
-  displayName: conditionWrapperSchemaV2.extract('displayName').messages({
-    'string.empty': 'Enter condition name'
-  })
-})
 
 export default [
   /**
@@ -299,10 +239,7 @@ export default [
 ]
 
 /**
- * @typedef {Partial<Omit<ConditionWrapperV2, 'items'>> & { action?: string, removeAction?: string, items?: Partial<ConditionDataV2>[] }} ConditionWrapperPayload
- */
-
-/**
- * @import { ConditionDataV2, ConditionWrapperV2 } from '@defra/forms-model'
+ * @import { ConditionWrapperV2 } from '@defra/forms-model'
  * @import { ServerRoute } from '@hapi/hapi'
+ * @import { ConditionWrapperPayload } from '~/src/models/forms/editor-v2/condition-helper.js'
  */
