+++ designer/server/src/views/forms/editor-v2/condition-check-changes.njk
@@ -2,7 +2,6 @@
 
 {% from "page-body/macro.njk" import appPageBody %}
 {% from "editor-card/macro.njk" import appEditorCard %}
-{% from "condition-editor/macro.njk" import appConditionEditor %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 
 {% set mainClasses = "govuk-main-wrapper--masthead" %}
